+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 
 Unreleased
 ==========
-
+* feat: MenuContent and MenuItem links open in sideframe (refactored icons)
 
 1.0.1 (2022-03-01)
 ===================